Add files via upload
Moved from ecen403 folder to DLPScanner. Also moved several old modules and scripts that are broken and/or unused into old_code. Major changes were made to cvstereo.py. Pointcloud generation no longer segfaults (hooray!), and lots of bugs were squashed. Also, pcd.py was added to be used in lieu of the pypcd library, which is not supported by Python 3. The modules usb.py and wifi.py were both modified to make use of it.